rough faker values for mocked response

+++ Lib/models/propertyDetails/Assessment.cs
@@ -3,7 +3,7 @@ namespace Lib.models.propertyDetails{
     public class Assessment {
         public string? Apn {get; set;}
         public int? Assessment_year {get; set;}
-        public int? Tax_amount {get; set;}
+        public decimal? Tax_amount {get; set;}
         public int? Tax_year {get; set;}
         public decimal? Total_assessed_value {get; set;}
     }
